server finish

+++ src/routers/seller.router.js
@@ -1,9 +1,9 @@
 const router = require('express').Router()
 const PostsModel = require('../models/posts.model')
 const UsersModel = require('../models/users.model')
-const { upload, validateUpload, verifyToken, verifyRoles, deleteFile, sendEmail } = require('../lib/functions')
+const ReservationsModel = require('../models/reservations.model')
+const { pictureUpload, deleteFileUpload, deleteMultiFile, deleteOneFile, upload, validateUpload, verifyToken, verifyRoles, sendEmail } = require('../lib/functions')
 const { msgNewPost } = require('../lib/msg')
-const { Post } = require('../lib/http')
 
 const uploadImgPost = upload('img', 500000, /png|jpg|jpeg/, 'posts_picture', 7)
 
@@ -14,6 +14,7 @@ router.get('/getPosts', verifyToken, verifyRoles(['Seller']), async (req, res) =
 })
 
 router.post('/newPost', verifyToken, verifyRoles(['Seller']), validateUpload(uploadImgPost), async (req, res) => {
+    let count = 0
     const { 
         title,
         subtitle,
@@ -23,6 +24,7 @@ router.post('/newPost', verifyToken, verifyRoles(['Seller']), validateUpload(upl
         currency,
         boatType,
         boatSize,
+        minHours,
         crew,
         city,
         marinaBeach,
@@ -32,16 +34,28 @@ router.post('/newPost', verifyToken, verifyRoles(['Seller']), validateUpload(upl
         includeDrinks,
         bathrooms,
         bedrooms,
-        kitchen
+        kitchen,
+        hasAirConditioning,
+        hasBluetoothSound,
+        hasLounge,
+        hasTV,
+        hasTowels,
+        hasDishes,
+        beers,
+        sodas,
+        hasIce,
+        tableWaters,
     } = req.body
     const { _id } = req.dataUser
     const files = req.files
     const imgUrls = []
     const imgPaths = []
-    files.map(({filename, path}) => { 
-        imgUrls.push(`${process.env.ApiUrl}assets/posts_picture/${filename}`)
-        imgPaths.push(path)
-    })
+
+    for(const file of files){
+        const { secure_url, public_id } = await pictureUpload(file.path, 500, 500, 'posts_picture')
+        imgUrls.push(secure_url)
+        imgPaths.push(public_id)
+    }
 
     const newPost = new PostsModel({
         idUser: _id,
@@ -55,6 +69,7 @@ router.post('/newPost', verifyToken, verifyRoles(['Seller']), validateUpload(upl
         imgPaths,
         boatType,
         boatSize: parseFloat(boatSize),
+        minHours: parseInt(minHours),
         crew: parseInt(crew),
         city,
         marinaBeach,
@@ -64,7 +79,17 @@ router.post('/newPost', verifyToken, verifyRoles(['Seller']), validateUpload(upl
         includeDrinks,
         bathrooms: parseInt(bathrooms),
         bedrooms: parseInt(bedrooms),
-        kitchen
+        kitchen,
+        hasAirConditioning,
+        hasBluetoothSound,
+        hasLounge,
+        hasTV,
+        hasTowels,
+        hasDishes,
+        beers: parseInt(beers),
+        sodas,
+        hasIce,
+        tableWaters: parseInt(tableWaters)
     })
 
     try {
@@ -75,7 +100,8 @@ router.post('/newPost', verifyToken, verifyRoles(['Seller']), validateUpload(upl
         await newPost.save()
         res.json({ server: 'postCreated'}).status(200)
     } catch(e) {
-        await deleteFile(files)
+        await deleteFileUpload(imgPaths)
+        await deleteMultiFile(files)
         res.json({ server: 'postNotCreated'}).status(200)
     }
 })
@@ -86,10 +112,11 @@ router.put('/updatePost/:idPost', verifyToken, verifyRoles(['Seller']), validate
     const files = req.files
     const imgUrls = []
     const imgPaths = []
-    files.map(({filename, path}) => { 
-        imgUrls.push(`${process.env.ApiUrl}assets/posts_picture/${filename}`)
-        imgPaths.push(path)
-    })
+    for(const file of files){
+        const { secure_url, public_id } = await pictureUpload(file.path, 500, 500, 'posts_picture')
+        imgUrls.push(secure_url)
+        imgPaths.push(public_id)
+    }
 
     const { 
         title,
@@ -100,6 +127,7 @@ router.put('/updatePost/:idPost', verifyToken, verifyRoles(['Seller']), validate
         currency,
         boatType,
         boatSize,
+        minHours,
         crew,
         city,
         marinaBeach,
@@ -109,12 +137,22 @@ router.put('/updatePost/:idPost', verifyToken, verifyRoles(['Seller']), validate
         includeDrinks,
         bathrooms,
         bedrooms,
-        kitchen
+        kitchen,
+        hasAirConditioning,
+        hasBluetoothSound,
+        hasLounge,
+        hasTV,
+        hasTowels,
+        hasDishes,
+        beers,
+        sodas,
+        hasIce,
+        tableWaters,
     } = req.body
 
     const post = await PostsModel.findOne({_id: idPost}).exec()
     if(post !== null && post.idUser === _id) {
-        const deletedFile = await deleteFile(post.imgPaths)
+        const deletedFile = await deleteFileUpload(post.imgPaths)
         if(deletedFile) {
             const updatePost = await PostsModel.updateOne({ _id: idPost }, {
                 title,
@@ -127,6 +165,7 @@ router.put('/updatePost/:idPost', verifyToken, verifyRoles(['Seller']), validate
                 imgPaths,
                 boatType,
                 boatSize: parseFloat(boatSize),
+                minHours: parseInt(minHours),
                 crew: parseInt(crew),
                 city,
                 marinaBeach,
@@ -136,21 +175,34 @@ router.put('/updatePost/:idPost', verifyToken, verifyRoles(['Seller']), validate
                 includeDrinks,
                 bathrooms: parseInt(bathrooms),
                 bedrooms: parseInt(bedrooms),
-                kitchen
+                kitchen,
+                hasAirConditioning,
+                hasBluetoothSound,
+                hasLounge,
+                hasTV,
+                hasTowels,
+                hasDishes,
+                beers: parseInt(beers),
+                sodas,
+                hasIce,
+                tableWaters: parseInt(tableWaters)
             }).exec()
-        
+
             if(updatePost.modifiedCount === 1) res.json({server: 'updatedPost'}) 
             else {
-                await deleteFile()
+                await deleteFileUpload(imgPaths)
+                await deleteMultiFile(imgPaths)
                 res.json({server: 'updatedNotPost'})
             }
         } else {
-            await deleteFile(imgPaths)
+            await deleteFileUpload(imgPaths)
+            await deleteMultiFile(imgPaths)
             res.json({server: 'updatedNotPost'})
         }
         
     } else {
-        await deleteFile(imgPaths)
+        await deleteFileUpload(imgPaths)
+        await deleteMultiFile(imgPaths)
         res.json({server: 'postNotExist'})
     }
 })
@@ -162,13 +214,16 @@ router.delete('/deletePost/:idPost', verifyToken, verifyRoles(['Seller']), async
     const post = await PostsModel.findOne({_id: idPost}).exec()
 
     if(post !== null && post.idUser === _id) {
-        const deletedFile = await deleteFile(post.imgPaths)
-        if(deletedFile) {
-            const deletePost = await PostsModel.deleteOne({ _id: idPost }).exec()
-        
-            if(deletePost.deletedCount === 1) res.json({server: 'deletedPost'}) 
-            else res.json({server: 'deletedNotPost'})
-        } else res.json({server: 'deletedNotPost'})
+        const boatIsReserved = await ReservationsModel.findOne({idPost}).exec()
+        if(boatIsReserved === null) {
+            const deletedFile = await deleteFileUpload(post.imgPaths)
+            if(deletedFile) {
+                const deletePost = await PostsModel.deleteOne({ _id: idPost }).exec()
+            
+                if(deletePost.deletedCount === 1) res.json({server: 'deletedPost'}) 
+                else res.json({server: 'deletedNotPost'})
+            } else res.json({server: 'deletedNotPost'})
+        } else res.json({server: 'postIsReservedItCannotBeDeleted'})
     } else res.json({server: 'postNotExist'})
 })
